requirements.txt
# =============================================================================
# Email Agent - Python Dependencies
# =============================================================================

# --- Google Cloud SDK ---
google-cloud-storage>=2.10.0,<3.0.0
google-cloud-language>=2.11.0,<3.0.0
google-cloud-vision>=3.4.0,<4.0.0
google-api-python-client>=2.90.0,<3.0.0
google-auth>=2.20.0,<3.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth-httplib2>=0.1.0,<1.0.0
oauth2client>=4.1.0,<5.0.0

# --- Web Framework (Cloud Run) ---
Flask>=2.3.0,<4.0.0
gunicorn>=21.0.0,<23.0.0
Werkzeug>=2.3.0,<4.0.0

# --- Cloud Functions Framework ---
functions-framework==3.4.0

# --- AI/ML - Gemini and Embeddings ---
google-generativeai>=0.7.0,<1.0.0
sentence-transformers>=2.2.0,<3.0.0

# --- Vector Database ---
chromadb>=0.4.0,<1.0.0

# --- LangChain Ecosystem ---
langchain-core>=0.1.0,<1.0.0
langchain-community>=0.0.20,<1.0.0
langchain>=0.1.0,<1.0.0
langchain-huggingface>=0.0.1,<1.0.0

# --- HTTP and API ---
requests>=2.31.0,<3.0.0

# --- Data Processing ---
python-dateutil>=2.8.0,<3.0.0
protobuf>=4.21.0,<5.0.0
typing-extensions>=4.5.0,<5.0.0
nltk>=3.8.0,<4.0.0
beautifulsoup4==4.12.2
tqdm>=4.65.0,<5.0.0

# --- Google Sheets Integration ---
gspread>=5.10.0,<6.0.0

# --- Input Validation and Sanitization ---
email-validator>=2.0.0,<3.0.0
bleach>=6.0.0,<7.0.0
python-magic>=0.4.27,<1.0.0

# --- Image Processing ---
Pillow>=10.0.0,<11.0.0

# --- Configuration ---
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0.0

# --- Performance Monitoring ---
psutil>=5.9.0,<6.0.0
deprecation==2.0.2

# --- Testing (Development Only) ---
pytest>=7.0.0,<8.0.0
pytest-mock>=3.10.0,<4.0.0
pytest-asyncio>=0.21.0,<1.0.0